refactor: consolidate mypage info
